+++ changelog.md
@@ -1,2 +1,2 @@
-Laseredstone 1.0.3+1.21.8
-- Lens is now correctly non-opaque
+Laseredstone 1.0.4+1.21.8
+- Optimize lasers checking for redstone power